requirements.txt
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
requests>=2.31.0
pyyaml>=6.0
